CI: refactor MicroVMBuilder class
In the old version this class would act as a helper
for constructing microvms with its internal fields always
changing to mirror the fields of the microvm. This would
make it impossible to call a destructor for the builder to clean
up resources.
This class now acts as true builder in the sense that it
now creates a common folder for all the microVMs
it creates and when it goes out of scope the folder
with all the resources can be deleted.
Also, this commit moves the VmNano and VmMicro classes under the
MicroVMBuilder.

Signed-off-by: Diana Popa <dpopa@amazon.com>