rate-limiter: increase limiter precision
The firecracker rate-limiter implementation leans more on efficiency
at the expense of accuracy.

It is a passive implementation driven by the consumer. Each time 'X'
tokens go through it, the rate-limiter:
 1. calculates how much time has passed since last operation
 2. replenishes its budget based on elapsed time above
 3. attempts to allow X if it fits the budget

Because we don't want to incur the cost of floating point operations
the operation in 'step 2' is a non-floating point division which
loses precision:
  // At each 'time_delta' nanoseconds the bucket should refill with:
  refill_amount = (time_delta * size) / (refill_time_ms * 1_000_000)

The smaller 'time_delta * size' and the more of these operations we
do, the higher our precision loss.

We can't control 'size' since it's user-provided input, but we can
slightly change our implementation to optimize on the other two
dimensions: we want bigger 'time_delta's and fewer division operations.

These are complementary and can be achieved by doing bucket replenish
operations only when needed (budget is low/empty) instead of every
time the limiter sees traffic.

This small change results in increased precision (it should be cheaper
too since we're doing fewer operations).

Signed-off-by: Adrian Catangiu <acatan@amazon.com>