tests: fix screen cleanup
The current way of terminating screen sessions does
not kill the underlying Firecracker processes.

Moreover, because the screen sessions are not killed,
and because `test_serial_io.py::test_serial_block` sends
SIGSTOP to screen, to stop capturing the output, the next
tests that are run with the jailer non-daemonized, will block
on `screen -ls` indefinitely. Probably because it leaves
the screen session in an invalid state.

Replaced this with a classic `kill -9`, that fixes the
above.

Signed-off-by: alindima <alindima@amazon.com>