rename seccompiler binary to seccompiler-bin
Once we upstream the seccompiler library to rust-vmm and use it
from there, the seccompiler binary in Firecracker will need a
different name, so that it doesn't create confusion.

As such, we rename the seccompiler binary to seccompiler-bin now,
so that we don't rename it later, avoiding a breaking change in naming.

The seccompiler library keeps its name and it will be used, under
the hood by the binary as well.

Signed-off-by: alindima <alindima@amazon.com>