seccompiler: rename arg_index and arg_type
In the JSON format, the `args` property maps to a vector
of SeccompConditions, which has fields for arg index and
arg type, among others.

Renamed the arg_index to index and arg_type to type, since
the naming was a little redundant.

This is the proposed version in the rust-vmm seccompiler
crate.

Signed-off-by: alindima <alindima@amazon.com>