fix metrics timestamp using monotonic clock
The utc_timestamp_ms field that is reported by the metrics
uses a CLOCK_MONOTONIC timestamp. The POSIX standard defines
the monotonic clock as a clock with a fixed but undefined
starting point.
For linux, the starting point is usually set at boot.

This commit replaces the clock with a CLOCK_REALTIME, which
gives the interval from the UTC UNIX epoch. It also updates
the changelog and adds a regression test.

Signed-off-by: alindima <alindima@amazon.com>