Add VM cleanup for tests using MicrovmBuilder
Adds a destructor for the VmInstance class that calls
the vm.kill() cleanup method.

We also need to replace the vm() getter method with the
vm proerty itself, because the python garbage collector is not
smart enough to not destruct VmInstance if we get a reference
to the vm using a getter.

Signed-off-by: alindima <alindima@amazon.com>