vm-memory: add test_utils module
This module exports a couple of test helpers for
creating the guest memory.

The first one is for creating the guest
memory without guard pages.

This is needed because the default `create_guest_memory`
uses MmapRegionBuilder::build_raw() for setting up the
memory with guard pages, which would error if the size is
not a multiple of the page size.
We need a test helper because We have unit tests which
need a custom memory size, not a multiple of the page size.

The second test helpers creates anonymous guest memory
objects, that do not need a FileOffset object.
It is just a little syntactic sugar that helps deduplicate
test code.

Signed-off-by: alindima <alindima@amazon.com>