[doc]: Update documentation regarding arp tables
If before snapshot there is an internenet connection the
guest will populate its arp caches and it will be saved
in memory snapshot file. If we delete the tap device and
recreate it the information will not be valid. The
interval of time when packets are lost is due to ARP
broadcast packet being sent to all the hosts on the
network requesting the physical hardware address for the
host. To avoid this issue the arp tables have to be dropped
after restore.

Signed-off-by: AlexandruCihodaru <cihodar@amazon.com>