jailer: add support for cgroupv2
This patch introduces support for configuring cgroups on
a cgroupv2 unified hierarchies by introducing a command
line argument --cgroup-version that allows a user to
choose which version of cgroup to use.
While cgroupv1 and cgroupv2 hierarchies can coexist in
a hybrid mode, in practice cgroupv1 hierarchies hijack
all controllers at boot time which leaves them unusable
for cgroupv2.
We only allow creation of cgroups into a single type of
hierarchy, not implementing support for hybrid mode.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>