[CI]: Test for mmap allow on api thread
When there is a put or patch request with a large header there will
be a mmap call from api_server thread with MAP_PRIVATE flag. It was
not allowed. Now we are able to send large buffers. This should be
completed by checking the size of the body and returing error when
the length is greater than a transhold.

Signed-off-by: AlexandruCihodaru <cihodar@amazon.com>
Signed-off-by: Luminita Voicu <lumivo@amazon.com>