split the boot args in regular args and init args
This split is needed because we're altering the kernel commandline
passed by Firecracker customers to add virtio device configuration (on
x86_64). The virtio config needs to be specified *before* the init.

This is not the ideal implementation of the fix, as it would make more
sense to have it in rust-vmm/linux-loader. Due to existing technical
debt implementing it directly in upstream is not straightforward.
See: https://github.com/rust-vmm/linux-loader/issues/92.

Fixes: https://github.com/firecracker-microvm/firecracker/issues/2709

Signed-off-by: Andreea Florescu <fandree@amazon.com>