enhance /snapshot/load API
Deprecate 'mem_file_path' field and add a new 'mem_backend'
optional object. At least one of 'mem_file_path' and `mem_backend`
fields are required and it is forbidden to specify both at the
same time.

`mem_backend` object contains:
- 'backend_type': required String parameter that can take either:
    - File
    - Uffd
as valid values.
- 'backend_path': required String parameter. Interpretation of
this field depends on the value of 'backend_type':
    - Path to file that contains the guest memory to be loaded
      if type is 'File',
    - Path to UDS where a custom page-fault handler process is
      listening and expecting a Uffd to be sent by Firecracker.
      The Uffd is used to handle Firecracker's guest memory page
      faults in this separate process.

If `mem_file_path` is specified instead of `mem_backend`, we
construct the `MemBackendConfig` object from the path specified,
through `mem_file_path`, with `File` as backend type.

Add plumbing for the new API, the Uffd path is only a stub for
now.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>
Signed-off-by: Luminita Voicu <lumivo@amazon.com>