add permanent /version HTTP endpoint
This endpoint, defined by a GET request on /version is a permanent
endpoint in our API starting with v1.0.0 which will never get removed.
It is especially important to define it in such a way that it will
stand the test of time without breaking changes made to it. We will
be able to add more fields to the JSON dictionary returned as a
response to this request, but we won't be able to change what we add.

Signed-off-by: George Pisaltu <gpl@amazon.com>