fix jailer --new-pid-ns tests
When starting the jailer in a new pid namespace, whether we use the
newpid_cloner or the screen approach, we will have problems.
1. The newpid_cloner creates a new pid namespace in which it runs
the jailer. The jailer creates a new pid namespace for the firecracker
process and kills itself. But the jailer is  the process with pid 1
inside the pid namespace created by the newpid_cloner, so when it dies
the entire namespace is cleaned and the firecracker process is killed.
2. Similarly, when using the screen command, the jailer will create a
new pid namespace for the firecracker process, then kill itself. After
the jailer process kills itself, the screen process gets killed as well
since it's not running any process inside anymore.

We fallback to a different branch that uses fork when offloading the pid
namespacing to the jailer.

Signed-off-by: Serban Iorga <seriorga@amazon.com>