CI: enforce dependency check
We have created a file at tests/framework/dependencies.txt
where we hold the current build dependencies. The test
fails if the number of dependencies is bigger and displays
the difference.

The test also fails if the expected dependencies are different
than the current ones. This mechanism ensures that the list of
dependencies stays up to date and that whenever a dependency
is removed we reset our dependency number.

Signed-off-by: Diana Popa <dpopa@amazon.com>