feat: add mmds content from file
This change allows adding of mmds from a supplied file when starting
Firecracker using `--metadata`. This brings it in line with the existing
`--config-file` in allowing you to configure Firecracker via files on
the command line which allows you to use no-api.

The test framework has been updated to expose the `--metadata` parameter
on the `MicroVM` class so that it can be used in tests.

Signed-off-by: Richard Case <richard@weave.works>