[io_uring] reorganize submission logic
Calling `submit_and_wait()` with the number of pending sqes wasn't a
reliable method to wait for the completion of all ops. Apart from the
pending sqes, we should also account for the available cqes, and also
since the io_uring ops execute on a separate thread, this numbers can
change while doing the calculation. In order to reliably understand what
is the total number of pending ops we need to keep an internal evidence.

Signed-off-by: Serban Iorga <seriorga@amazon.com>