add runbook for changing the Firecracker API
This commit adds the runbook for changing the API. This is a guide for
contributors which should be followed whenever one needs to make an
API change in Firecracker.

The runbook details all the steps necessary to change the API and also
comes with examples for doing so. This is part of our effort towards
having a robust API support policy.

Signed-off-by: George Pisaltu <gpl@amazon.com>