recreate tag of greatest release
GitHub tags releases based on the creation date of their
associated tag. When we publish a patch release of an older
release, GitHub marks that release with the `latest` label.
Their definition of `latest` is `most recent`. However, our
definition of `latest` is `the one with the greatest version`, so
we want to have `latest` tag on the greatest released Firecracker
version, for better visibility for our customers. The way to
achieve this is to recreate the tag of the greatest release
version, after creating the tag associated with the patch release
we want to publish. This way, the tag for the release with the
greatest version is also the most recent tag created and GitHub
will assign it the `latest` label.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>