[CI]: Fix snapshot perf test
We are using a global list to store the scratch drives info.
In function that makes the setup for block scaling test we
were changing the type of that global variable from list to
zip, thus in test function the list was seen as empty. In python
changing the type of a global variable might cause issue. Returned
a list from the function used to intialize the global to solve this
issue. Also added catch for AssertionError in run_exercise.

Signed-off-by: AlexandruCihodaru <cihodar@amazon.com>