seccomp: move mmap variant to correct thread
During this commit: a3f5d47b063599e2eb73a116e537988de53b2929,
the mmap call was mistakenly added to the vmm thread, as opposed
to the API thread to which it was intended.

Also fix the regression test. There was a bug in it, which lead
to this issue being hidden.

Signed-off-by: alindima <alindima@amazon.com>