tests: undo skipping of tests
We were previously skipping some snapshotting-related
integration_tests tests due to outdated S3 artifacts.
Now that snapshotting support is merged, I uploaded
fresh v1.0 artifacts to S3 and now all tests are passing.

This commit reverts the change and adds a temporary S3 folder
for the io_uring feature branch.

A separate folder is needed due to the changes on the main
branch on the vsock_id field and the incompatible snapshot
states.

When the feature branch is merged in main we can go remove this
temporary folder in S3.

Signed-off-by: alindima <alindima@amazon.com>