if metadata file: set version; populate MMDS
When using `--metadata` cmdline argument, we populate MMDS
contents from file after setting MMDS version. This is because
in the unlikely case of an error during MMDS initialization
(if V2 cannot be set successfully), MMDS object will be None.
This is expected: it means that MMDS is not available to provide
V2 functionality, but V1 can still be configured with no problems.

But the MMDS contents are gone. In case the user does not care
about V2 and configures mmds-version to V1 through the config
file, data store will be empty. If however, we populate MMDS after
configuring the version to V1, the data will be successfully
placed inside the data store.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>