add `network_interfaces` field to `/mmds/config`
Enhance `/mmds/config` to contain a list of the network interface
IDs capable of forwarding packets to MMDS. Network interface IDs
mentioned must be valid at the time of this request. The net
device model will reply to HTTP GET requests sent to the MMDS
address via the interfaces mentioned. In this case, both ARP
requests and TCP segments heading to `ipv4_address` are
intercepted by the device model, and do not reach the associated
TAP device.

This implementation makes the `allow-mmds-requests` field set
when creating a net device redundant. Will remove the
`allow-mmds-requests` field in a subsequent commit.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>