mmds: store expiry in ms instead of seconds
This commit changes expiration time's unit of measure from seconds
to milliseconds. When computing expiry in seconds, if the request
to create a new token arrives at time X,999 seconds with a ttl=1s,
the token's expiry is computed to be X + 1 and the token ends up
having a lifetime of 0,0001 seconds instead of 1 second.

Computing expiry based on current time in milliseconds instead of
seconds provides greater precision.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>