fix race in HostEchoServer.exit()
There is a race between the exit() function closing the
host vsock UDS and run() function which polls the socket
using select().
If the select() runs right after the socket is closed but
before the thread is signalled to stop, we get a BadFd errno,
causing a false test failure.

This is the root cause of intermittent failures in
check_guest_connections affecting tests like
test_vsock.py and test_snapshot_basic.py

Signed-off-by: alindima <alindima@amazon.com>