api/machine_config: fix PUT and PATCH behavior
Refactor machine_config update path to make it more consistent.
This fixes several issues:
* Setting the cpu_template and smt value through PUT and PATCH is not
  allowed for aarch64 but can be done without error by starting a
  microvm from a JSON file.
* If machine_config PUT request is issued without the track_dirty_pages
  and the cpu_template parameters set, track_dirty_pages would be set
  to false but the cpu_template would remain unchanged.
* There is no way to reset the cpu_template through PUT or PATCH once it
  has been set with either of these methods.
* Patching track_dirty_pages was not allowed on aarch64 even though it
  is a pre-boot request.

This patch splits the actual machine configuration in 2 structs: one
used for keeping the actual configured values and one used only for
updates. Since Firecracker starts with default values for the machine
configuration, all changes (either PUT or PATCH) are unified through an
update path.
Also having 2 distinct structs allows for validation of parameters early
in the deserialization of the structs, and eliminates the need for
handling unwraps for the parameters in the VMM.
Since the parameters are validated in the deserializer, we can use the
same code for PUT, PATCH and starting from a json file.

Also added more details in the swagger file to better explain the
behaviour of PUT.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>