vmm: fix drop logic for Vmm
Before this fix, if creating the Vmm object would fail during
a snapshot restore, after restoring the vcpus, Firecracker
would panic at the assert line inside Vmm's drop(). This is
because when trying to propagate the error, the Vmm object
would be dropped, without clearing out the contents of
`vcpu_handles` (without joining the vcpu threads).

By explicitly calling `stop()` insinde the dropping logic of the
Vmm object, we ensure that the Vmm is always ready to be teared
down when the dropping logic kicks in.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>