ACPI Shutdown (fixed hardware method) implemented
A port IO device implemented for ACPI registers X_PM1a_EVT_BLK and
X_PM1a_CNT_BLK required for a shutdown using fixed hardware method.
In order to perform the shutdown, we need to set a couple of bits
from these registers according to the spec and trigger an interupt
in order to make OSPM read these registers and generate the ACPI
event.

Signed-off-by: Traistaru Andrei Cristian <atc@amazon.com>