api: move MMDS actions to VMM
Move the Get, Put and Patch MMDS API requests
to be serviced from the VMM thread, like every other
Firecracker request.
This brings consistency with the rest of the code.

It will also make it easier to remove the lazy_static usage for
the data store, since it will be one less place to refactor.
We will be able to have one single entity responsible for the
data store, the VMM, making the code more easily readable.

Signed-off-by: alindima <alindima@amazon.com>