[mmds]: Check ethertype before the address
The speculative check of destination address could have been
bypassed with a special crafted packet. If a IPv4 packet had
the wrong destination address but at the offset where the
address is in an ARP packet there is the correct destination
address then the speculative check would be bypassed.

Remve speculative check and check the destination address at the
correct offeset after the ethertype is checked.

Signed-off-by: AlexandruCihodaru <cihodar@amazon.com>
Suggested-by: Radu Iliescu <riliescu@amazon.com>