mmds: remove lazy_static usage
Refactor the code using mmds so that we may remove the
lazy_static MMDS global variable.

The Mmds data store is now kept on the VmResources object,
accessible from the Vmm thread, as well as a shared reference
across all Net devices configured to service Mmds requests.

This is a big step towards removing global state from
Firecracker and will enable us to not create the Mmds data store
unless the user actually configured it.

Signed-off-by: alindima <alindima@amazon.com>