vmm: increase recv_timeout limit
This hardcoded limit is used when the vmm waits on vcpu
responses after broadcasting a Resume/Pause/Save/Restore
event.

Increased this limit from 1s to 30s so that valid use cases
having a high host load are not hindered. We keep a timeout
so that an unresponsive vcpu would be detected.

Signed-off-by: alindima <alindima@amazon.com>