uffd: don't send fc's PID to PF handler over UDS
The page fault handler process requires FC's PID in order
to be able to notify FC of any crashes/exists. The way
we initially intended to make the handler process aware
of FC's PID was to have FC sent its PID explicitly through
the UDS. But there is a problem with this approach:

Firecracker's PID depends on the way the process was started.
When normally started through the jailer, Firecracker will
see its global PID. However, if using `--new-pid-ns` flag
when running the jailer, the jailer will spawn FC in a new
PID namespace, which would cause Firecracker to not be
aware of its PID and see it as being 1.

Moreover, when not using `--new-pid-ns` flag, we recommend
to the users to spawn the jailer process into a new PID
namespace when starting it.

This is why sending FC's PID from whithin the FC process is
not reliable.

To tackle this, we no longer send FC's PID explicitly,
but advise the page fault handler process to use
`getsockopt` call with `SO_PEERCRED` option in order
to fetch the global Firecracker PID at the moment of
connecting.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>