net: use only `prepare_kick()` to raise queue irqs
Instead of keeping track whether we have consumed any descriptors in a
separate variable, rely on `prepare_kick()` to find out whether we
should enable guest notifications.

This will probably induce a performance hit with guests that don't
support notification suppression, but it simplifies the code in the
common case.

Signed-off-by: Babis Chalios <bchalios@amazon.es>