add --mmds-size-limit parameter
Previously, we used to piggy-back on the
`--http-api-max-payload-size` for setting the mmds size
limit.
There was actually a bug in the code that prevented Firecracker
from limiting this data store when populating it from a file.

This commit introduces a separate parameter for the data store
limit, which defaults to the api limit if left unconfigured
(for backwards compatibility).
It also fixes the mentioned bug and adds a regression test, as
well as adding documentation updates.

Signed-off-by: alindima <alindima@amazon.com>