api: fix race in api connecting logic
We have been observing intermittent failures for
test_start_with_metadata_default_limit. This comes
from a race possibility when the microVM is not
succesfuly built and thus there is no previous
blocking communication between the API thread and
Firecracker.
If that happens, we may run the risk of Firecracker trying
to connect to the socket path to trigger a graceful shutdown
of the API thread before the socket was actually created.

Signed-off-by: Diana Popa <dpopa@amazon.com>