CI: fix sporadic fail in serial_after_snapshot
The reason we were seeing intermittent fails was due
to the test logic not reaching a deterministic point in
time before snapshotting the vm. As such, sometimes there
would be serial content after resuming from snapshot triggered
by logging in the user but sometimes, the user login would be much
faster and appear before the snapshot.

Signed-off-by: Diana Popa <dpopa@amazon.com>