tests/snap: make balloon checks more reliable
When checking the functionality of the balloon device, we use an
application to fault in some memory inside the guest, then try to
reclaim some memory back using the balloon.

Make the balloon check function fault in and reclaim a more significant
amount of memory such that the test does not get affected by other
processes in the guest that are maybe using memory temporarily.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>