net: create `IoVecBuffer` type for vectored writes
The standard library handles vectored writes and reads using the
`IoSlice` and `IoSliceMut` respectively. We introduce a new type,
`IoVecBuffer` which essentially is a vector of `IoSlice` and is able to
be instantiated from a `DescriptorChain`.

`IoVecBuffer` provides us the necessary bits for using `write_vectored`
to transmit packets in the tap.

Signed-off-by: Babis Chalios <bchalios@amazon.es>