Rely on env variable for kcov run detect
There is logic inside the unit tests which is not compatible
when run with kcov. A good example are the tests looking into
signal handling logic.

Since, kcov already intercepts SIGSYS on x86_64 and SIGILL
on aarch64 and thus would not reach Firecracker which in turn
would lead to some assertion not being the expected ones.

So, what we do here, we add an environment variable that signals
whether the unit tests are being run with kcov. Based on that, we
do not look at logic involving SIGILL on arm or SIGSYS on x86_64.

Signed-off-by: Diana Popa <dpopa@amazon.com>
(cherry picked from commit 0269d1cc3201311ca3740a2043638345fa30e2a9)