cpu templates: add T2S templates
The T2 template lets KVM populate the ARCH_CAPABILITIES MSR by using a
copy of the Host information. We sometimes want to force the information
contained in that MSR in order to make it look like a less capable
processor WRT speculative execution vulnerabilities and mitigations.
This is helpful when we want to snapshot a uVM on a newer processor and
resume on an older processor.
In this case, the software should use only the capabilities that are
available in both the newer and older host.

Create a new CPU template called T2S which is similar to the T2 template
but also sets additional information in the ARCH_CAPABILITIES MSR.
We basically set the MSR to expose to the guest a CPU that has the
security capabilities of a Skylake CPU, regardless of the host.
Also set some bits that were missing from the T2 CPUID config.

Adding the MSR information to the cpuid crate is not ideal based on
the current implementation. In the end the whole cpuid crate should
be refactored to remove the templates implementation into another crate.
The assumption that CPUID is the only place where information about the
processor is stored proves to be false so we need to extend the
templates to include at least the MSRs.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>