tests/api: fix test_api_version test
The Firecracker version returned by the API should always be the one
returned by `git describe --dirty`.
This is done to identify Firecracker versions when debugging.
The version is embedded in the Firecracker binary by the `build.rs`
script by running `git describe --dirty`.
The API version test was using different parameters than `build.rs` to
get the version and was also discarding the hash, `wip` and `dirty`
strings from the version which is not what we want to do.

Fixed the test by removing the sanitization part and the `--abbrev=0`
parameter.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>