inconsistent instance naming for ARM baselines
The latest perf baselines for ARM contain 2 entries for
ARM: m6g.metal and m6gd.metal. That is because the parse_baselines
tool will insert a m6gd.metal one.

Replace m6gd.metal with m6g.metal (due to the fact that this is the
latest entry) and fix the parse_baselines tool.

Signed-off-by: Diana Popa <dpopa@amazon.com>