Comply with TOML string requirements
Drops string handling from pre-commit entirely.
rustfmt --config option can't be used with the parameters
'as is' (string values as unquoted). Instead use
rustfmt --config-path that takes the whole fmt.toml file.

To be able to use --config-path requires the nightly toolchain
to be added to the docker dev container as well.

This change additionally drops the license-template-path
formatting parameter as it is no longer supported by the
rust toolchain. See github discussion:
https://github.com/rust-lang/rustfmt/issues/3352

Signed-off-by: Matthew Schlebusch <schlebus@amazon.com>