acpi: add Virtual Machine Generation ID support
Add a VMGenID ACPI device which maintains a 128-bits generation ID and
exposes it in system memory inside the guest. The purpose of the device
is to update the ID every time the VM is restored from a snapshot and
notify the guest. The guest can use the notification and the value of
the generation ID in order to re-adjust itself to the new environment.

At the moment, Linux uses the generation ID to re-seed its PRNG every
time it changes.

Additionally, add a Generic Event Device (GED), which is an ACPI
mechanism for handling interrupts directed to ACPI devices and
performing actions as a result of the interrupt event. GED is used to
deliver to VMGenID device a notification whenever the generation ID
changes.

Signed-off-by: Babis Chalios <bchalios@amazon.es>