vmm: handle uniformly GSIs and memory of a guest
Right now we have one pair of (IdAllocator, AddressAllocator) per device
manager (mmio and legacy for x86, mmio for aarch64) and we separate the
resources (GSIs and guest memory) these handle.

With ACPI we will add new types of devices that will need as well GSIs
and guest memory, which would mean that with the current way of things
we would need to create yet another pair of allocators.

This fragments the resources (we should divide the GSIs used by ACPI and
MMIO devices for example) and keeps information about them fragmented.

This commit introduces a global resource manager for GSIs and guest
memory. Its scope is the whole VM rather than type of devices, so we can
avoid the aforementioned fragmentation.

Signed-off-by: Babis Chalios <bchalios@amazon.es>