tests: cpu features: add msr restore test
This adds a 2-part test to verify that MSRs retain their values after
restoring from a snapshot.
The first part of the test makes modifications to MSRs, dumps MSRs
to a file and takes a snapshot.
The second part of the test restores from the snapshot and compares
the two sets of the MSR values.

Currently two tests of part one are added:
 - "empty": does not make any MSR modifications prior to taking snapshot
 - "some": add some MSR modificaitons.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.co.uk>