[CI] devtool: Use ramdisk for TMPDIR
Currently some tests are failing when running `devtool test` using a
ramdisk, because they are trying to create hard links between `/srv`
(which is inside the ramdisk) and `/tmp` (which is not). However,
cross-device hard links are not allowed.

When running with a ramdisk make `$TMPDIR` point to a directory inside
it.

Signed-off-by: Babis Chalios <bchalios@amazon.es>