virtio-queue: panic on invalid avail ring index
The number of descriptor chain heads to process should always be smaller
or equal to the queue size, as the driver should never ask the VMM to
process a available ring entry more than once. Checking and reporting
such incorrect driver behavior can prevent potential hanging and
Denial-of-Service from happening on the VMM side.

Issue reported in rust-vmm/vm-virtio and fixed in
https://github.com/rust-vmm/vm-virtio/pull/196.

Signed-off-by: Diana Popa <dpopa@amazon.com>
Co-authored-by: Bo Chen <chen.bo@intel.com>