Fix `test_restrictions` io_uring test
For unallowed operations we were testing that adding a bad opcode in the
ring would result in an `Other` error when trying to pop from the ring.
The `ErrorKind` changed from `Other` to `Uncategorized` which is an
unstable variant which we can't match on.

So, instead of checking for the particular variant we just check for
error now.

Signed-off-by: Babis Chalios <bchalios@amazon.es>