tests: cpu features: add cpuid snapshot test
The 2-part test checks if the CPUID retains its value after a VM is
restored from a snapshot.
It dumps CPUID before taking snapshot, dumps CPUID after restoring
from the snapshot and then compares the two.

pylint's too-many-lines is disabled in the test file as there is no
good way to split the file at the moment.  As it grows further,
there should be an opportunity to reenable the check, for example,
by refactoring.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>