devtool: parse minor kernel version when building
At the moment, we pass devtool a kernel configuration when building. The
script parses the kernel version from the configuration file in order to
fetch the proper tarball with kernel source to build. However, at the
moment it only parses up to major version so, for example, a
configuration produced from 4.14.174 would result in 4.14 being built.

This commit changes the regex used for parsing in order to parse the
minor version as well.

Signed-off-by: Babis Chalios <bchalios@amazon.es>