fix(tests): avoid naming conflicts in tap devices
Many tests create a tap0 before moving it to its own network namespace.
When running tests with multiple workers in parallel, there can be
naming conflicts if tap0 is already created.

This patch avoids it by creating the tap device with a random name
first, then moving it to the netns with the correct name.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>