tests: upgrade snap resume baseline infra
The purpose of this change is to align the test framework with the one
in the main branch (as of 32c7f1e220bf46f988301b915bb5f995772a31aa) to
allow snap restore baseline collection in firecracker-v1.1 and
comparison with the baselines collected in main.
The major difference compensated is the baseline config file schema.

The following has been updated:
 - test_snapshot_restore_performance.py
 - snap restore baseline config files
 - parse_baseline script.

Style of some Python files needed to be updated to satisfy the linter
used before switching to black after v1.1.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>