Fix cross compiling firecracker
When cross compiling rust binaries, some components nevertheless have
to be compiled for the host architecture. These are proc-macros and
build  scripts. In vmm(build.rs) we use cargo run to execute
seccompiler-bin.  However, we pass the --target flag, meaning
seccompiler-bin is not  built for the architecture it is being run
on. This commit removes the --target flag, meaning seccompiler-bin
will always be built for the host architecture on which it is run

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>